move middleware into its own file for organization
